* encrypt.c (encrypt_is): Cast arguments to printf to match format
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@13384 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/appl/telnet/libtelnet/encrypt.c
@@ -563,7 +563,8 @@ encrypt_is(data, cnt)
 	} else {
 		ret = (*ep->is)(data, cnt);
 		if (encrypt_debug_mode)
-			printf("(*ep->is)(%x, %d) returned %s(%d)\n", data, cnt,
+			printf("(*ep->is)(%x, %d) returned %s(%d)\n", 
+			        (unsigned int) data, cnt,
 				(ret < 0) ? "FAIL " :
 				(ret == 0) ? "SUCCESS " : "MORE_TO_DO ", ret);
 	}
@@ -608,7 +609,7 @@ encrypt_reply(data, cnt)
 		ret = (*ep->reply)(data, cnt);
 		if (encrypt_debug_mode)
 			printf("(*ep->reply)(%x, %d) returned %s(%d)\n",
-				data, cnt,
+				(unsigned int) data, cnt,
 				(ret < 0) ? "FAIL " :
 				(ret == 0) ? "SUCCESS " : "MORE_TO_DO ", ret);
 	}
